Refactor Constant propagation into analysis and transformer
